+++ docs/make.jl
@@ -53,6 +53,7 @@ makedocs(
       "ACPFlow Runner" => "reference_acpflow.md",
       "Network Model" => "reference_network.md",
       "Rectangular Power Flow" => "reference_powerflow_rectangular.md",
+      "DC Power Flow" => "reference_powerflow_dc.md",
       "Import and Export" => "reference_import_export.md",
       "State Estimation and Measurements" => "reference_state_estimation.md",
     ],