+++ test/runtests.jl
@@ -32,6 +32,7 @@ function include_fast_tests()
   include("test_matpower_metadata.jl")
   include("test_api.jl")
   include("test_webui.jl")
+  include("test_dc_powerflow.jl")
 end
 
 function include_extended_tests()
@@ -66,6 +67,7 @@ function run_fast_profile_tests()
     ("programmatic_api", () -> run_entry(:run_api_fast_tests)),
     ("webui", () -> run_entry(:run_webui_fast_tests)),
     ("state_estimation", () -> run_entry(:run_state_estimation_tests)),
+    ("dc_powerflow", () -> run_entry(:run_dc_powerflow_tests)),
     ("controls", () -> begin
       run_entry(:run_voltage_dependent_control_tests)
       run_entry(:run_transformer_phase_shift_tests)